refactor: split internal into config + cli packages
Extracts configuration loading, validation, and the Duration type into
a new internal/config/ package (package config), leaving CLI/lifecycle
in internal/ (package internal).

- internal/config/config.go — Configuration struct, ReadConf, factory
  methods (GetChecks, GetDelays, GetDownAction, GetStatServerConfig)
  env-var expansion, probe creation
- internal/config/config_validate.go — field-level validation
- internal/config/duration.go — Duration type (moved from internal/)
- internal/config/config_test.go, config_validate_test.go,
  duration_test.go — corresponding tests
- internal/cmd.go — imports config package, SetupLoop uses
  config.ReadConf and config.Configuration
- internal/cmd_test.go — defines testConfigDir locally
- Deleted internal/conf.go, conf_validate.go, duration.go and their
  test files (now in internal/config/)
