+++ internal/config/config_validate.go
@@ -1,4 +1,4 @@
-package internal
+package config
 
 import (
 	"errors"
