+++ internal/config/config_validate_test.go
@@ -1,4 +1,4 @@
-package internal
+package config
 
 import (
 	"os"
